sniper-main/requirements.txt
requests
tabulate
python-dotenv
pandas
